test: move portal tests
