+++ gravitee-apim-e2e/api-test/src/portal/mapi-v1/portal-business-error-apis.spec.ts
@@ -16,16 +16,16 @@
 import { afterAll, beforeAll, describe, test } from '@jest/globals';
 
 import { forManagementAsAdminUser, forPortalAsAdminUser, forPortalAsAppUser } from '@gravitee/utils/configuration';
-import { PortalApi as PortalManagementApi } from '@gravitee/management-webclient-sdk/src/lib/apis/PortalApi';
-import { APIsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIsApi';
+import { PortalApi as PortalManagementApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/PortalApi';
+import { APIsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIsApi';
 import { ApisFaker } from '@gravitee/fixtures/management/ApisFaker';
-import { UpdateApiEntityFromJSON } from '@gravitee/management-webclient-sdk/src/lib/models/UpdateApiEntity';
-import { ApiApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/ApiApi';
+import { UpdateApiEntityFromJSON } from '../../../../lib/management-webclient-sdk/src/lib/models/UpdateApiEntity';
+import { ApiApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/ApiApi';
 import { fail, notFound } from '@lib/jest-utils';
-import { ApiEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApiEntity';
-import { RatingEntity } from '@gravitee/management-webclient-sdk/src/lib/models/RatingEntity';
+import { ApiEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiEntity';
+import { RatingEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/RatingEntity';
 import { PortalApiFaker } from '@gravitee/fixtures/management/PortalApiFaker';
-import { ApiResponse } from '../../../lib/portal-webclient-sdk/src/lib';
+import { ApiResponse } from '../../../../lib/portal-webclient-sdk/src/lib';
 
 const orgId = 'DEFAULT';
 const envId = 'DEFAULT';
