+++ gravitee-apim-e2e/api-test/src/portal/mapi-v1/portal-business-error-auth.spec.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { describe } from '@jest/globals';
-import { AuthenticationApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/AuthenticationApi';
+import { AuthenticationApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/AuthenticationApi';
 import { forPortalAsAdminUser } from '@gravitee/utils/configuration';
 import { notFound } from '@lib/jest-utils';
