+++ gravitee-apim-e2e/api-test/src/portal/mapi-v1/portal-business-error-subscriptions.spec.ts
@@ -16,28 +16,28 @@
 import { afterAll, beforeAll, describe, expect, test } from '@jest/globals';
 
 import { forManagementAsAdminUser, forPortalAsAdminUser, forPortalAsSimpleUser } from '@gravitee/utils/configuration';
-import { APIsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIsApi';
+import { APIsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIsApi';
 import { ApisFaker } from '@gravitee/fixtures/management/ApisFaker';
 import { fail } from '@lib/jest-utils';
-import { ApiEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApiEntity';
-import { ApiLifecycleState } from '@gravitee/management-webclient-sdk/src/lib/models/ApiLifecycleState';
-import { Visibility } from '@gravitee/management-webclient-sdk/src/lib/models/Visibility';
+import { ApiEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiEntity';
+import { ApiLifecycleState } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiLifecycleState';
+import { Visibility } from '../../../../lib/management-webclient-sdk/src/lib/models/Visibility';
 import { PlansFaker } from '@gravitee/fixtures/management/PlansFaker';
 import { ApplicationsFaker } from '@gravitee/fixtures/management/ApplicationsFaker';
-import { APIPlansApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIPlansApi';
-import { ApplicationsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/ApplicationsApi';
-import { PlanEntity } from '@gravitee/management-webclient-sdk/src/lib/models/PlanEntity';
-import { ApplicationEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApplicationEntity';
-import { SubscriptionApi as PortalSubscriptionApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/SubscriptionApi';
-import { APISubscriptionsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APISubscriptionsApi';
-import { Subscription } from '@gravitee/management-webclient-sdk/src/lib/models/Subscription';
-import { ApplicationSubscriptionsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/ApplicationSubscriptionsApi';
-import { ApiKeyEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApiKeyEntity';
-import { PlanSecurityType } from '@gravitee/management-webclient-sdk/src/lib/models/PlanSecurityType';
-import { PlanValidationType } from '@gravitee/management-webclient-sdk/src/lib/models/PlanValidationType';
-import { SubscriptionStatus } from '@gravitee/management-webclient-sdk/src/lib/models/SubscriptionStatus';
+import { APIPlansApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIPlansApi';
+import { ApplicationsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/ApplicationsApi';
+import { PlanEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanEntity';
+import { ApplicationEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApplicationEntity';
+import { SubscriptionApi as PortalSubscriptionApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/SubscriptionApi';
+import { APISubscriptionsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APISubscriptionsApi';
+import { Subscription } from '../../../../lib/management-webclient-sdk/src/lib/models/Subscription';
+import { ApplicationSubscriptionsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/ApplicationSubscriptionsApi';
+import { ApiKeyEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiKeyEntity';
+import { PlanSecurityType } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanSecurityType';
+import { PlanValidationType } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanValidationType';
+import { SubscriptionStatus } from '../../../../lib/management-webclient-sdk/src/lib/models/SubscriptionStatus';
 import faker from '@faker-js/faker';
-import { UpdateApiEntityFromJSON } from '@gravitee/management-webclient-sdk/src/lib/models/UpdateApiEntity';
+import { UpdateApiEntityFromJSON } from '../../../../lib/management-webclient-sdk/src/lib/models/UpdateApiEntity';
 
 const orgId = 'DEFAULT';
 const envId = 'DEFAULT';
