+++ gravitee-apim-e2e/api-test/src/portal/mapi-v1/portal-business-error-users.spec.ts
@@ -15,13 +15,13 @@
  */
 import { afterAll, beforeAll, describe, test } from '@jest/globals';
 import { forManagementAsAdminUser, forPortalAsAdminUser, forPortalAsSimpleUser } from '@gravitee/utils/configuration';
-import { UsersApi as UsersManagementApi } from '@gravitee/management-webclient-sdk/src/lib/apis/UsersApi';
+import { UsersApi as UsersManagementApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/UsersApi';
 import { UsersFaker } from '@gravitee/fixtures/management/UsersFaker';
-import { UserEntity } from '@gravitee/management-webclient-sdk/src/lib/models/UserEntity';
+import { UserEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/UserEntity';
 import { fail, notFound, unauthorized } from '@lib/jest-utils';
-import { UsersApi as UsersPortalApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/UsersApi';
-import { UserApi as UserPortalApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/UserApi';
-import { PortalApi as PortalManagementApi } from '@gravitee/management-webclient-sdk/src/lib/apis/PortalApi';
+import { UsersApi as UsersPortalApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/UsersApi';
+import { UserApi as UserPortalApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/UserApi';
+import { PortalApi as PortalManagementApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/PortalApi';
 
 const portalManagementApi = new PortalManagementApi(forManagementAsAdminUser());
 const usersManagementApiAsAdmin = new UsersManagementApi(forManagementAsAdminUser());
